Criação do Dashboard

+++ Models/ProdutoModel.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public string? Nome { get; set; }
         public string? Tipo { get; set; }
-        public decimal? PrecoUnitario { get; set; }
+        public float? PrecoUnitario { get; set; }
     }
 }
